fix(git): address review feedback — API consistency, unquote paths, restore comments
- Unify checkoutBranch to use GitCommand.run directly (matching checkoutBranchAsync)
- Add unquoteGitPath to parseIgnoredOutput for C-quoted paths with spaces/Unicode
- Move GitFileStatus.color to separate GitFileStatus+Color.swift (Foundation-only models)
- Restore key comments explaining why in GitStatusProvider
- Add test for ignored paths containing spaces
